requirements.txt
fastapi 
uvicorn 
torchaudio 
pyannote.audio 
openai-whisper 
pydantic
huggingface_hub
pymongo
python-dotenv
webrtcvad
httpx
groq